Refactor builders into separate files
