+++ tct/src/internal/frontier/tier/builder.rs
@@ -1,0 +1,151 @@
+use super::*;
+
+pub struct Builder<Item: Built + Focus>
+where
+    Item::Complete: Built,
+{
+    index: u64,
+    global_position: u64,
+    inner: Option<InnerBuilder<Item>>,
+}
+
+enum InnerBuilder<Item: Built + Focus>
+where
+    Item::Complete: Built,
+{
+    Frontier(<Nested<Item> as Built>::Builder),
+    Complete(<<Nested<Item> as Focus>::Complete as Built>::Builder),
+}
+
+impl<Item: Focus + Built + Height> Built for Tier<Item>
+where
+    Item::Complete: Built,
+{
+    type Builder = Builder<Item>;
+
+    fn build(global_position: u64, index: u64) -> Self::Builder {
+        Builder {
+            index,
+            global_position,
+            inner: None,
+        }
+    }
+}
+
+impl<Item: Built + Focus> Build for Builder<Item>
+where
+    Item::Complete: Built,
+{
+    type Output = Tier<Item>;
+
+    fn go(mut self, instruction: Instruction) -> Result<IResult<Self>, HitBottom<Self>> {
+        use {IResult::*, Instruction::*};
+
+        if let Some(inner) = self.inner {
+            // If we're already building something, pass the instruction along to the inside:
+            match inner {
+                InnerBuilder::Frontier(builder) => match builder.go(instruction) {
+                    Err(HitBottom(builder)) => {
+                        self.inner = Some(InnerBuilder::Frontier(builder));
+                        Err(HitBottom(self))
+                    }
+                    Ok(Incomplete(builder)) => {
+                        self.inner = Some(InnerBuilder::Frontier(builder));
+                        Ok(Incomplete(self))
+                    }
+                    Ok(Complete(frontier)) => Ok(Complete(Tier {
+                        inner: Inner::Frontier(Box::new(frontier)),
+                    })),
+                },
+                InnerBuilder::Complete(builder) => match builder.go(instruction) {
+                    Err(HitBottom(builder)) => {
+                        self.inner = Some(InnerBuilder::Complete(builder));
+                        Err(HitBottom(self))
+                    }
+                    Ok(Incomplete(builder)) => {
+                        self.inner = Some(InnerBuilder::Complete(builder));
+                        Ok(Incomplete(self))
+                    }
+                    Ok(Complete(complete)) => Ok(Complete(Tier {
+                        inner: Inner::Complete(Box::new(complete)),
+                    })),
+                },
+            }
+        } else if let Leaf { here } = instruction {
+            // If we're not yet building anything and we receive our first instruction as a `Leaf`,
+            // then immediately return a completed hashed tier
+            Ok(Complete(Tier {
+                inner: Inner::Hash(Hash::new(here)),
+            }))
+        } else {
+            // Otherwise, our instruction is to builder a witnessed tier, so set that up
+            // and follow the instruction:
+
+            // In which we do some math to determine whether or not the node is on the frontier...
+
+            // The height of this tier
+            let height = <Self::Output as Height>::Height::HEIGHT;
+            // The number of positions each index increment corresponds to
+            let stride = 4u64.pow(height.into());
+            // The position of the zeroth child of this tier
+            let start_position = stride * self.index;
+            // The position after the last child of this tier
+            let end_position_non_inclusive = (start_position + stride).min(4u64.pow(24) - 1);
+            // Whether the node is on the frontier
+            let frontier =
+                (start_position..end_position_non_inclusive).contains(&self.global_position);
+
+            self.inner = if frontier {
+                Some(InnerBuilder::Frontier(Item::build(
+                    self.global_position,
+                    self.index,
+                )))
+            } else {
+                Some(InnerBuilder::Complete(Item::Complete::build(
+                    self.global_position,
+                    self.index,
+                )))
+            };
+
+            // Now that we've set up the inside, use the instruction to proceed
+            self.go(instruction)
+        }
+    }
+
+    fn is_started(&self) -> bool {
+        self.inner.is_some()
+    }
+
+    fn index(&self) -> u64 {
+        if let Some(inner) = &self.inner {
+            match inner {
+                InnerBuilder::Frontier(frontier) => frontier.index(),
+                InnerBuilder::Complete(complete) => complete.index(),
+            }
+        } else {
+            self.index
+        }
+    }
+
+    fn height(&self) -> u8 {
+        if let Some(inner) = &self.inner {
+            match inner {
+                InnerBuilder::Frontier(frontier) => frontier.height(),
+                InnerBuilder::Complete(complete) => complete.height(),
+            }
+        } else {
+            <Self::Output as Height>::Height::HEIGHT
+        }
+    }
+
+    fn min_required(&self) -> usize {
+        if let Some(inner) = &self.inner {
+            match inner {
+                InnerBuilder::Frontier(frontier) => frontier.min_required(),
+                InnerBuilder::Complete(complete) => complete.min_required(),
+            }
+        } else {
+            1
+        }
+    }
+}
